user Api and some changes in like component

+++ LikeAndCommentService/likecomment/src/main/java/com/example/likecomment/model/LikePostModel.java
@@ -17,7 +17,7 @@ public class LikePostModel {
     @Transient
     private UserSendPostModel userLike;
 
-    private Long userLikeId;
+    private Integer userLikeId;
     private boolean likeUserActive;
 
     private boolean isActive;
